Create Misc module with various small test stuff
which doesnt really belon anywhere particurly
