+++ hs-bindgen/hs-bindgen.cabal
@@ -137,7 +137,7 @@ test-suite golden
   type:           exitcode-stdio-1.0
   main-is:        golden.hs
   hs-source-dirs: tests
-  other-modules:  Orphans TH TastyGolden
+  other-modules:  Orphans TH TastyGolden Misc
   build-depends:
       -- Internal dependencies
     , hs-bindgen
